flashbloc/flashblocproject/requirements.txt
asgiref==3.5.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.0.3
django-cors-headers==4.0.0
django-extensions==3.2.1
django-formtools==2.4.1
django-otp==1.2.0
django-phonenumber-field==6.4.0
django-postgres-dropdb==2020.12.3
django-two-factor-auth==1.15.2
Django-Verify-Email==2.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
idna==3.4
phonenumbers==8.13.11
pip==23.0.1
postgres==4.0
psycopg2==2.9.3
psycopg2-pool==1.1
PyGreSQL==5.2.4
PyJWT==2.7.0
pypng==0.20220715.0
pytz==2021.3
qrcode==7.4.2
requests==2.31.0
setuptools==58.0.4
sqlparse==0.4.2
typing_extensions==4.6.2
tzdata==2021.5
urllib3==2.0.3
vercel==0.2.1
wheel==0.37.1
wincertstore==0.2
